More test refactoring
